+++ system/src/Grav/Framework/Flex/FlexForm.php
@@ -394,18 +394,6 @@ class FlexForm implements FlexObjectFormInterface, \JsonSerializable
     }
 
     /**
-     * Implements \Serializable::unserialize().
-     *
-     * @param string $data
-     */
-    public function unserialize($data): void
-    {
-        $data = unserialize($data, ['allowed_classes' => [FlexObject::class]]);
-
-        $this->doUnserialize($data);
-    }
-
-    /**
      * @param string $name
      * @return mixed|null
      */
@@ -456,6 +444,14 @@ class FlexForm implements FlexObjectFormInterface, \JsonSerializable
      */
     public function __unset($name)
     {
+    }
+
+    /**
+     * @return array|bool
+     */
+    protected function getUnserializeAllowedClasses()
+    {
+        return [FlexObject::class];
     }
 
     /**