+++ system/src/Grav/Framework/Object/Base/ObjectTrait.php
@@ -9,6 +9,8 @@
 
 namespace Grav\Framework\Object\Base;
 
+use Grav\Framework\Compat\Serializable;
+
 /**
  * Object trait.
  *
@@ -16,6 +18,8 @@ namespace Grav\Framework\Object\Base;
  */
 trait ObjectTrait
 {
+    use Serializable;
+
     /** @var string */
     protected static $type;
 
@@ -119,25 +123,23 @@ trait ObjectTrait
     }
 
     /**
-     * Implements Serializable interface.
-     *
-     * @return string
+     * @return array
      */
-    public function serialize()
+    final public function __serialize(): array
     {
-        return serialize($this->doSerialize());
+        return $this->doSerialize();
     }
 
     /**
-     * @param string $serialized
+     * @param array $data
+     * @return void
      */
-    public function unserialize($serialized)
+    final public function __unserialize(array $data): void
     {
-        $data = unserialize($serialized);
-
         if (method_exists($this, 'initObjectProperties')) {
             $this->initObjectProperties();
         }
+
         $this->doUnserialize($data);
     }